+++ src/tt_api/src/components/js/Navbar.js
@@ -1,6 +1,6 @@
 import React, {useState, useEffect} from 'react';
 import { Link } from 'react-router-dom';
-import './Navbar.css';
+import '../css/Navbar.css';
 
 function Navbar() {
     const [click, setClick] = useState(false);
