+++ src/tt_api/src/pages/js/Dev.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import '../../App.css';
-import Footer from '../Footer';
+import Footer from './Footer';
 
 function Developers () {
     return (
