+++ src/tt_api/src/pages/js/Samples.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import '../../App.css';
-import Footer from '../Footer';
-import './Samples.css';
+import '../css/Samples.css';
+import Footer from './Footer';
 
 function Samples () {
     return (
